reduced everything to one page
