--- pages/not_found_404.py
@@ -1,7 +1,0 @@
-from dash import html
-import dash
-
-dash.register_page(__name__, path="/404")
-
-
-layout = html.H1("Custom 404")
